+++ tests/pure/es.aggregate-error.js
@@ -1,5 +1,6 @@
 import AggregateError from 'core-js-pure/es/aggregate-error';
 import Symbol from 'core-js-pure/es/symbol';
+import toString from 'core-js-pure/es/object/to-string';
 
 QUnit.test('AggregateError', assert => {
   assert.isFunction(AggregateError);
@@ -14,4 +15,5 @@ QUnit.test('AggregateError', assert => {
   assert.same(AggregateError([1]).message, '');
   assert.deepEqual(AggregateError([1, 2, 3]).errors, [1, 2, 3]);
   assert.throws(() => AggregateError([1], Symbol()), 'throws on symbol as a message');
+  assert.same(toString(AggregateError([1])), '[object Error]', 'Object#toString');
 });